refactor: align oaks impact school quote prop types with sanity
