+++ src/components/GenericPagesComponents/OaksImpactSchoolQuote/OaksImpactSchoolQuote.test.tsx
@@ -1,4 +1,4 @@
-import { fixtureData } from "./OaksImpactSchoolQuote.fixtures";
+import { fixtureData } from "../OaksImpactSchoolQuotesSection/OaksImpactSchoolQuotesSection.fixtures";
 
 import { OaksImpactSchoolQuote } from ".";
 
@@ -6,10 +6,12 @@ import { renderWithProvidersByName } from "@/__tests__/__helpers__/renderWithPro
 
 const render = renderWithProvidersByName(["oakTheme"]);
 
+const fixture = fixtureData.cards[0]!;
+
 describe("OaksImpactSchoolQuote", () => {
   it("renders correctly", () => {
     const { baseElement, getByRole } = render(
-      <OaksImpactSchoolQuote {...fixtureData} />,
+      <OaksImpactSchoolQuote {...fixture} />,
     );
     expect(baseElement).toMatchSnapshot();
     expect(getByRole("heading")).toHaveTextContent("Ormiston Academies Trust");
@@ -19,7 +21,7 @@ describe("OaksImpactSchoolQuote", () => {
 
   it("renders as='li'", () => {
     const { container } = render(
-      <OaksImpactSchoolQuote {...fixtureData} as="li" />,
+      <OaksImpactSchoolQuote {...fixture} as="li" />,
     );
     expect(container).toMatchSnapshot();
     expect(container.firstElementChild?.tagName).toBe("LI");
