+++ src/components/GenericPagesComponents/OaksImpactSchoolQuotesSection/OaksImpactSchoolQuotesSection.fixtures.ts
@@ -1,0 +1,99 @@
+import { OaksImpactSchoolQuotesSectionProps } from ".";
+
+export const fixtureData: OaksImpactSchoolQuotesSectionProps = {
+  heading: "How schools and MATs are embedding Oak’s curriculum",
+  cards: [
+    {
+      logo: {
+        altText: "TEST_ALT",
+        asset: {
+          _id: "image-b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264-jpg",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264.jpg`,
+        },
+      },
+      summary:
+        "Ormiston Academies Trust supports pupils in 45 schools across secondary, primary, alternative provision and special schools.",
+      quote: {
+        text: "In Ormiston, we value our partnership with Oak, through which we have strengthened the R.E. curriculum in our schools, whilst contributing to a high-quality national resource for all.\nOrmiston primary schools are using science and humanities resources to strengthen their curricular offer, and are benefitting from the high quality curriculum design, and opportunities to work in partnership with colleagues at Oak.",
+        attribution: "Tom Rees",
+        role: "CEO",
+        organisation: "Ormiston Academies Trust",
+      },
+      headshot: {
+        asset: {
+          _id: "image-d40d4c63b117e248ec11b995951aa738969531aa-54x54-png",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/d40d4c63b117e248ec11b995951aa738969531aa-54x54.png`,
+        },
+      },
+    },
+    {
+      logo: {
+        altText: "TEST_ALT",
+        asset: {
+          _id: "image-b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264-jpg",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264.jpg`,
+        },
+      },
+      summary:
+        "Ormiston Academies Trust supports pupils in 45 schools across secondary, primary, alternative provision and special schools.",
+      quote: {
+        text: "In Ormiston, we value our partnership with Oak, through which we have strengthened the R.E. curriculum in our schools, whilst contributing to a high-quality national resource for all.\nOrmiston primary schools are using science and humanities resources to strengthen their curricular offer, and are benefitting from the high quality curriculum design, and opportunities to work in partnership with colleagues at Oak.",
+        attribution: "Tom Rees",
+        role: "CEO",
+        organisation: "Lift Schools",
+      },
+      headshot: {
+        asset: {
+          _id: "image-d40d4c63b117e248ec11b995951aa738969531aa-54x54-png",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/d40d4c63b117e248ec11b995951aa738969531aa-54x54.png`,
+        },
+      },
+    },
+    {
+      logo: {
+        altText: "TEST_ALT",
+        asset: {
+          _id: "image-b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264-jpg",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264.jpg`,
+        },
+      },
+      summary:
+        "Ormiston Academies Trust supports pupils in 45 schools across secondary, primary, alternative provision and special schools.",
+      quote: {
+        text: "In Ormiston, we value our partnership with Oak, through which we have strengthened the R.E. curriculum in our schools, whilst contributing to a high-quality national resource for all.\nOrmiston primary schools are using science and humanities resources to strengthen their curricular offer, and are benefitting from the high quality curriculum design, and opportunities to work in partnership with colleagues at Oak.",
+        attribution: "Tom Rees",
+        role: "CEO",
+        organisation: "Primary School",
+      },
+      headshot: {
+        asset: {
+          _id: "image-d40d4c63b117e248ec11b995951aa738969531aa-54x54-png",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/d40d4c63b117e248ec11b995951aa738969531aa-54x54.png`,
+        },
+      },
+    },
+    {
+      logo: {
+        altText: "TEST_ALT",
+        asset: {
+          _id: "image-b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264-jpg",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/b81ee19a35baa3192360a210fda34cc9b21f4fd6-5824x3264.jpg`,
+        },
+      },
+      summary:
+        "Ormiston Academies Trust supports pupils in 45 schools across secondary, primary, alternative provision and special schools.",
+      quote: {
+        text: "In Ormiston, we value our partnership with Oak, through which we have strengthened the R.E. curriculum in our schools, whilst contributing to a high-quality national resource for all.\nOrmiston primary schools are using science and humanities resources to strengthen their curricular offer, and are benefitting from the high quality curriculum design, and opportunities to work in partnership with colleagues at Oak.",
+        attribution: "Tom Rees",
+        role: "CEO",
+        organisation: "Secondary School",
+      },
+      headshot: {
+        asset: {
+          _id: "image-d40d4c63b117e248ec11b995951aa738969531aa-54x54-png",
+          url: `https://sanity-asset-cdn.thenational.academy/images/cuvjke51/production/d40d4c63b117e248ec11b995951aa738969531aa-54x54.png`,
+        },
+      },
+    },
+  ],
+};
